An example using regex was added
Check for valid phone number
