+++ PlayingWithRe/library/__init__.py
@@ -1,0 +1,1 @@
+from .re_functions import replace_using_re
